refactor(interpreter): extract keyword-plist parsing from mcp.ts
`parsePlist`/`keyName` were private to mcp.ts, but the compression
built-ins need the same `:key value` calling convention. Importing them
from mcp.ts would drag the @modelcontextprotocol/sdk dependency into
every interp that loads compression — including the LSP's prelude-only
fallback and the test helpers — so they move to their own module.

Adds `plistOptions`, a strict variant that rejects an unrecognised key.
MCP tool calls must stay lenient (their keys come from a remote schema),
but for a built-in the keys are fixed and a silently-ignored typo is
worse than an error: an agent writing (view x :ofset 400) has to be
told, not handed the default window as if it had asked for it.
